// prisma/generated/zod/inputTypeSchemas/RefundLogUpdateInputSchema.ts
import { Prisma } from '@prisma/client';

import { z } from 'zod';
import { StringFieldUpdateOperationsInputSchema } from './StringFieldUpdateOperationsInputSchema';
import { isValidDecimalInput } from './isValidDecimalInput';
import { DecimalJsLikeSchema } from './DecimalJsLikeSchema';
import { DecimalFieldUpdateOperationsInputSchema } from './DecimalFieldUpdateOperationsInputSchema';
import { RefundStatusSchema } from './RefundStatusSchema';
import { EnumRefundStatusFieldUpdateOperationsInputSchema } from './EnumRefundStatusFieldUpdateOperationsInputSchema';
import { NullableStringFieldUpdateOperationsInputSchema } from './NullableStringFieldUpdateOperationsInputSchema';
import { NullableDateTimeFieldUpdateOperationsInputSchema } from './NullableDateTimeFieldUpdateOperationsInputSchema';
import { DateTimeFieldUpdateOperationsInputSchema } from './DateTimeFieldUpdateOperationsInputSchema';
import { RefundRequestUpdateOneRequiredWithoutRefundLogNestedInputSchema } from './RefundRequestUpdateOneRequiredWithoutRefundLogNestedInputSchema';
import { PaymentActionsUpdateOneRequiredWithoutRefundLogNestedInputSchema } from './PaymentActionsUpdateOneRequiredWithoutRefundLogNestedInputSchema';

export const RefundLogUpdateInputSchema: z.ZodType<Prisma.RefundLogUpdateInput> = z.object({
  id: z.union([ z.string().cuid(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  amount: z.union([ z.union([z.number(),z.string(),z.instanceof(Prisma.Decimal),DecimalJsLikeSchema,]).refine((v) => isValidDecimalInput(v), { message: 'Must be a Decimal' }),z.lazy(() => DecimalFieldUpdateOperationsInputSchema) ]).optional(),
  currency: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  provider: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  providerRefundId: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  status: z.union([ z.lazy(() => RefundStatusSchema),z.lazy(() => EnumRefundStatusFieldUpdateOperationsInputSchema) ]).optional(),
  errorMessage: z.union([ z.string(),z.lazy(() => NullableStringFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  processedAt: z.union([ z.coerce.date(),z.lazy(() => NullableDateTimeFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  createdAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  updatedAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  refundRequest: z.lazy(() => RefundRequestUpdateOneRequiredWithoutRefundLogNestedInputSchema).optional(),
  paymentTransaction: z.lazy(() => PaymentActionsUpdateOneRequiredWithoutRefundLogNestedInputSchema).optional()
}).strict();

export default RefundLogUpdateInputSchema;